chore: extracts expander plugin to multiple modules
