+++ packages/css/src/plugins/expand-shorthands/__tests__/background.test.tsx
@@ -1,0 +1,36 @@
+import postcss from 'postcss';
+import { propertyExpander } from '../../property-expander';
+
+const transform = (css: TemplateStringsArray) => {
+  const result = postcss([propertyExpander]).process(css[0], {
+    from: undefined,
+  });
+
+  return result.css;
+};
+
+describe('property expander', () => {
+  it('should expand background single', () => {
+    const result = transform`
+      background: red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            background-color: red;
+          "
+    `);
+  });
+
+  it('should do nothing when background is complex', () => {
+    const result = transform`
+      background: radial-gradient(crimson, skyblue);
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            background: radial-gradient(crimson, skyblue);
+          "
+    `);
+  });
+});
