+++ packages/css/src/plugins/expand-shorthands/__tests__/outline.test.tsx
@@ -1,0 +1,179 @@
+import postcss from 'postcss';
+import { propertyExpander } from '../../property-expander';
+
+const transform = (css: TemplateStringsArray) => {
+  const result = postcss([propertyExpander]).process(css[0], {
+    from: undefined,
+  });
+
+  return result.css;
+};
+
+describe('property expander', () => {
+  it('should expand outline single color', () => {
+    const result = transform`
+      outline: red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: red;
+            outline-style: initial;
+            outline-width: initial;
+          "
+    `);
+  });
+
+  it('should expand outline single style', () => {
+    const result = transform`
+      outline: solid;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: initial;
+            outline-style: solid;
+            outline-width: initial;
+          "
+    `);
+  });
+
+  it('should expand outline single width', () => {
+    const result = transform`
+      outline: 10px;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: initial;
+            outline-style: initial;
+            outline-width: 10px;
+          "
+    `);
+  });
+
+  it('should expand outline single width named', () => {
+    const result = transform`
+      outline: thin;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: initial;
+            outline-style: initial;
+            outline-width: thin;
+          "
+    `);
+  });
+
+  it('should expand outline double', () => {
+    const result = transform`
+      outline: red solid;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: red;
+            outline-style: solid;
+            outline-width: initial;
+          "
+    `);
+  });
+
+  it('should expand outline double reversed', () => {
+    const result = transform`
+      outline: solid red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: red;
+            outline-style: solid;
+            outline-width: initial;
+          "
+    `);
+  });
+
+  it('should expand outline triple', () => {
+    const result = transform`
+      outline: red solid 2px;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: red;
+            outline-style: solid;
+            outline-width: 2px;
+          "
+    `);
+  });
+
+  it('should expand outline triple reversed', () => {
+    const result = transform`
+      outline: solid 2px red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            outline-color: red;
+            outline-style: solid;
+            outline-width: 2px;
+          "
+    `);
+  });
+
+  it('should remove outline single invalid', () => {
+    const result = transform`
+      outline: asd;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+
+  it('should remove outline double invalid', () => {
+    const result = transform`
+      outline: asd solid;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+
+  it('should remove outline triple invalid', () => {
+    const result = transform`
+      outline: 10px solid asd;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+
+  it('should remove outline double invalid', () => {
+    const result = transform`
+      outline: 10px 10px;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+
+  it('should remove outline triple invalid', () => {
+    const result = transform`
+      outline: 10px solid 10px;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+});
