+++ packages/css/src/plugins/expand-shorthands/__tests__/overflow.test.tsx
@@ -1,0 +1,38 @@
+import postcss from 'postcss';
+import { propertyExpander } from '../../property-expander';
+
+const transform = (css: TemplateStringsArray) => {
+  const result = postcss([propertyExpander]).process(css[0], {
+    from: undefined,
+  });
+
+  return result.css;
+};
+
+describe('property expander', () => {
+  it('should expand overflow single', () => {
+    const result = transform`
+      overflow: hidden;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            overflow-x: hidden;
+            overflow-y: hidden;
+          "
+    `);
+  });
+
+  it('should expand overflow double', () => {
+    const result = transform`
+      overflow: auto hidden;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            overflow-x: auto;
+            overflow-y: hidden;
+          "
+    `);
+  });
+});
