+++ packages/css/src/plugins/expand-shorthands/__tests__/text-decoration.test.tsx
@@ -1,0 +1,79 @@
+import postcss from 'postcss';
+import { propertyExpander } from '../../property-expander';
+
+const transform = (css: TemplateStringsArray) => {
+  const result = postcss([propertyExpander]).process(css[0], {
+    from: undefined,
+  });
+
+  return result.css;
+};
+
+describe('property expander', () => {
+  it('should expand text decoration single', () => {
+    const result = transform`
+      text-decoration: underline;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            text-decoration-color: initial;
+            text-decoration-line: underline;
+            text-decoration-style: initial;
+          "
+    `);
+  });
+
+  it('should expand text decoration line double', () => {
+    const result = transform`
+      text-decoration: underline overline;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            text-decoration-color: initial;
+            text-decoration-line: overline underline;
+            text-decoration-style: initial;
+          "
+    `);
+  });
+
+  it('should remove text decoration invalid double', () => {
+    const result = transform`
+      text-decoration: underline underline;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+          "
+    `);
+  });
+
+  it('should expand text decoration double', () => {
+    const result = transform`
+      text-decoration: underline red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            text-decoration-color: red;
+            text-decoration-line: underline;
+            text-decoration-style: initial;
+          "
+    `);
+  });
+
+  it('should expand text decoration triple', () => {
+    const result = transform`
+      text-decoration: underline wavy red;
+    `;
+
+    expect(result).toMatchInlineSnapshot(`
+      "
+            text-decoration-color: red;
+            text-decoration-line: underline;
+            text-decoration-style: wavy;
+          "
+    `);
+  });
+});
