+++ packages/css/src/plugins/expand-shorthands/types.tsx
@@ -1,0 +1,4 @@
+import { Declaration } from 'postcss';
+import { Root as ValuesRoot } from 'postcss-values-parser';
+
+export type ConversionFunction = (node: Declaration, value: ValuesRoot) => Declaration[];
